v2/requirements.txt
fastapi
uvicorn
opencv-python
opencv-python-headless
numpy
Pillow
scikit-image